StudIA-main/requirements.txt
setuptools>=65.5.0
asgiref==3.7.2
Django==5.0.2
django-tenants==3.6.1
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
openpyxl==3.1.2
packaging==24.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv==1.0.0